refactor(web): extract shared media label mapping
